@@ -1,9 +1,8 @@
 module basic-go
 
-go 1.21.6
+go 1.23
 
 require (
-	github.com/DATA-DOG/go-sqlmock v1.5.0
 	github.com/IBM/sarama v1.43.1
 	github.com/alibabacloud-go/darabonba-openapi v0.2.1
 	github.com/alibabacloud-go/dysmsapi-20170525/v2 v2.0.18
