@@ -12,110 +12,91 @@ import (
 )
 
 var (
-	//go:embed lua/interative_incr_cnt.lua
+	//go:embed lua/incr_cnt.lua
 	luaIncrCnt string
 )
 
-const (
-	fieldReadCnt    = "read_cnt"
-	fieldCollectCnt = "collect_cnt"
-	fieldLikeCnt    = "like_cnt"
-)
+const fieldReadCnt = "read_cnt"
+const fieldLikeCnt = "like_cnt"
+const fieldCollectCnt = "collect_cnt"
 
-//go:generate mockgen -source=./interactive.go -package=cachemocks -destination=mocks/interactive.mock.go InteractiveCache
 type InteractiveCache interface {
-
-	// IncrReadCntIfPresent 如果在缓存中有对应的数据，就 +1
-	IncrReadCntIfPresent(ctx context.Context,
-		biz string, bizId int64) error
-	IncrLikeCntIfPresent(ctx context.Context,
-		biz string, bizId int64) error
-	DecrLikeCntIfPresent(ctx context.Context,
-		biz string, bizId int64) error
-	IncrCollectCntIfPresent(ctx context.Context, biz string, bizId int64) error
-	// Get 查询缓存中数据
-	Get(ctx context.Context, biz string, bizId int64) (domain.Interactive, error)
-	Set(ctx context.Context, biz string, bizId int64, intr domain.Interactive) error
+	IncrReadCntIfPresent(ctx context.Context, biz string, bizId int64) error
+	IncrLikeCntIfPresent(ctx context.Context, biz string, id int64) error
+	DecrLikeCntIfPresent(ctx context.Context, biz string, id int64) error
+	IncrCollectCntIfPresent(ctx context.Context, biz string, id int64) error
+	Get(ctx context.Context, biz string, id int64) (domain.Interactive, error)
+	Set(ctx context.Context, biz string, bizId int64, res domain.Interactive) error
 }
 
-type RedisInteractiveCache struct {
-	client     redis.Cmdable
-	expiration time.Duration
+type InteractiveRedisCache struct {
+	client redis.Cmdable
 }
 
-func (r *RedisInteractiveCache) IncrCollectCntIfPresent(ctx context.Context,
-	biz string, bizId int64) error {
-	return r.client.Eval(ctx, luaIncrCnt,
-		[]string{r.key(biz, bizId)},
-		fieldCollectCnt, 1).Err()
-}
-
-func (r *RedisInteractiveCache) IncrReadCntIfPresent(ctx context.Context,
-	biz string, bizId int64) error {
-	return r.client.Eval(ctx, luaIncrCnt,
-		[]string{r.key(biz, bizId)},
-		fieldReadCnt, 1).Err()
-}
-
-func (r *RedisInteractiveCache) IncrLikeCntIfPresent(ctx context.Context,
-	biz string, bizId int64) error {
-	return r.client.Eval(ctx, luaIncrCnt,
-		[]string{r.key(biz, bizId)},
-		fieldLikeCnt, 1).Err()
-}
-
-func (r *RedisInteractiveCache) DecrLikeCntIfPresent(ctx context.Context,
-	biz string, bizId int64) error {
-	return r.client.Eval(ctx, luaIncrCnt,
-		[]string{r.key(biz, bizId)},
-		fieldLikeCnt, -1).Err()
-}
-
-func (r *RedisInteractiveCache) Get(ctx context.Context,
-	biz string, bizId int64) (domain.Interactive, error) {
-	// 直接使用 HMGet，即便缓存中没有对应的 key，也不会返回 error
-	data, err := r.client.HGetAll(ctx, r.key(biz, bizId)).Result()
-	if err != nil {
-		return domain.Interactive{}, err
+func NewInteractiveRedisCache(client redis.Cmdable) InteractiveCache {
+	return &InteractiveRedisCache{
+		client: client,
 	}
-
-	if len(data) == 0 {
-		// 缓存不存在
-		return domain.Interactive{}, ErrKeyNotExist
-	}
-
-	// 理论上来说，这里不可能有 error
-	collectCnt, _ := strconv.ParseInt(data[fieldCollectCnt], 10, 64)
-	likeCnt, _ := strconv.ParseInt(data[fieldLikeCnt], 10, 64)
-	readCnt, _ := strconv.ParseInt(data[fieldReadCnt], 10, 64)
-
-	return domain.Interactive{
-		// 懒惰的写法
-		BizId:      bizId,
-		CollectCnt: collectCnt,
-		LikeCnt:    likeCnt,
-		ReadCnt:    readCnt,
-	}, err
 }
 
-func (r *RedisInteractiveCache) Set(ctx context.Context, biz string, bizId int64, intr domain.Interactive) error {
-	key := r.key(biz, bizId)
-	err := r.client.HMSet(ctx, key,
-		fieldLikeCnt, intr.LikeCnt,
-		fieldCollectCnt, intr.CollectCnt,
-		fieldReadCnt, intr.ReadCnt).Err()
+func (i *InteractiveRedisCache) Set(ctx context.Context,
+	biz string, bizId int64,
+	res domain.Interactive) error {
+	key := i.key(biz, bizId)
+	err := i.client.HSet(ctx, key, fieldCollectCnt, res.CollectCnt,
+		fieldReadCnt, res.ReadCnt,
+		fieldLikeCnt, res.LikeCnt,
+	).Err()
 	if err != nil {
 		return err
 	}
-	return r.client.Expire(ctx, key, time.Minute*15).Err()
+	return i.client.Expire(ctx, key, time.Minute*15).Err()
 }
 
-func (r *RedisInteractiveCache) key(biz string, bizId int64) string {
-	return fmt.Sprintf("interactive:%s:%d", biz, bizId)
-}
-
-func NewRedisInteractiveCache(client redis.Cmdable) InteractiveCache {
-	return &RedisInteractiveCache{
-		client: client,
+func (i *InteractiveRedisCache) Get(ctx context.Context, biz string, id int64) (domain.Interactive, error) {
+	key := i.key(biz, id)
+	res, err := i.client.HGetAll(ctx, key).Result()
+	if err != nil {
+		return domain.Interactive{}, err
 	}
+	if len(res) == 0 {
+		return domain.Interactive{}, ErrKeyNotExist
+	}
+	var intr domain.Interactive
+	intr.BizId = id
+	// 这边是可以忽略错误的
+	intr.CollectCnt, _ = strconv.ParseInt(res[fieldCollectCnt], 10, 64)
+	intr.LikeCnt, _ = strconv.ParseInt(res[fieldLikeCnt], 10, 64)
+	intr.ReadCnt, _ = strconv.ParseInt(res[fieldReadCnt], 10, 64)
+	return intr, nil
+}
+
+func (i *InteractiveRedisCache) IncrCollectCntIfPresent(ctx context.Context,
+	biz string, id int64) error {
+	key := i.key(biz, id)
+	return i.client.Eval(ctx, luaIncrCnt, []string{key}, fieldCollectCnt, 1).Err()
+}
+
+func (i *InteractiveRedisCache) IncrLikeCntIfPresent(ctx context.Context,
+	biz string, bizId int64) error {
+	key := i.key(biz, bizId)
+	return i.client.Eval(ctx, luaIncrCnt, []string{key}, fieldLikeCnt, 1).Err()
+}
+
+func (i *InteractiveRedisCache) DecrLikeCntIfPresent(ctx context.Context,
+	biz string, bizId int64) error {
+	key := i.key(biz, bizId)
+	return i.client.Eval(ctx, luaIncrCnt, []string{key}, fieldLikeCnt, -1).Err()
+}
+
+func (i *InteractiveRedisCache) IncrReadCntIfPresent(ctx context.Context,
+	biz string, bizId int64) error {
+	key := i.key(biz, bizId)
+	// 不是特别需要处理 res
+	//res, err := i.client.Eval(ctx, luaIncrCnt, []string{key}, fieldReadCnt, 1).Int()
+	return i.client.Eval(ctx, luaIncrCnt, []string{key}, fieldReadCnt, 1).Err()
+}
+
+func (i *InteractiveRedisCache) key(biz string, bizId int64) string {
+	return fmt.Sprintf("interactive:%s:%d", biz, bizId)
 }
