@@ -1,6 +1,6 @@
 package domain
 
 type WechatInfo struct {
-	OpenID  string
-	UnionID string
+	UnionId string
+	OpenId  string
 }
