@@ -1,119 +1,112 @@
 package cache
 
 import (
+	"basic-go/webook/internal/domain"
 	"context"
 	"encoding/json"
 	"fmt"
 	"time"
 
-	"basic-go/webook/internal/domain"
-
 	"github.com/redis/go-redis/v9"
 )
 
+//go:generate mockgen -source=./article.go -package=cachemocks -destination=./mocks/article.mock.go ArticleCache
 type ArticleCache interface {
-	// GetFirstPage 只缓存第第一页的数据
-	// 并且不缓存整个 Content
-	GetFirstPage(ctx context.Context, author int64) ([]domain.Article, error)
-	SetFirstPage(ctx context.Context, author int64, arts []domain.Article) error
-	DelFirstPage(ctx context.Context, author int64) error
-
-	Set(ctx context.Context, art domain.Article) error
+	GetFirstPage(ctx context.Context, uid int64) ([]domain.Article, error)
+	SetFirstPage(ctx context.Context, uid int64, res []domain.Article) error
+	DelFirstPage(ctx context.Context, uid int64) error
 	Get(ctx context.Context, id int64) (domain.Article, error)
-
-	// SetPub 正常来说，创作者和读者的 Redis 集群要分开，因为读者是一个核心中的核心
-	SetPub(ctx context.Context, article domain.Article) error
+	Set(ctx context.Context, art domain.Article) error
 	GetPub(ctx context.Context, id int64) (domain.Article, error)
+	SetPub(ctx context.Context, res domain.Article) error
 }
 
-type RedisArticleCache struct {
+type ArticleRedisCache struct {
 	client redis.Cmdable
 }
 
-func NewRedisArticleCache(client redis.Cmdable) ArticleCache {
-	return &RedisArticleCache{
+func (a *ArticleRedisCache) DelFirstPage(ctx context.Context, uid int64) error {
+	return a.client.Del(ctx, a.firstKey(uid)).Err()
+}
+
+func (a *ArticleRedisCache) Get(ctx context.Context, id int64) (domain.Article, error) {
+	val, err := a.client.Get(ctx, a.key(id)).Bytes()
+	if err != nil {
+		return domain.Article{}, err
+	}
+	var res domain.Article
+	err = json.Unmarshal(val, &res)
+	return res, err
+}
+
+func (a *ArticleRedisCache) Set(ctx context.Context, art domain.Article) error {
+	// JSON 序列化大部分场景，都不会引起性能问题
+	// 万一有性能问题怎么办？
+	// 换一个序列化方法，我推荐 protobuf
+	// 如果不想用 protobuf，可以考虑别的，例如 gob(Go Object Binary)
+	val, err := json.Marshal(art)
+	if err != nil {
+		return err
+	}
+	return a.client.Set(ctx, a.key(art.Id), val, time.Minute*10).Err()
+}
+
+func (a *ArticleRedisCache) GetPub(ctx context.Context, id int64) (domain.Article, error) {
+	val, err := a.client.Get(ctx, a.pubKey(id)).Bytes()
+	if err != nil {
+		return domain.Article{}, err
+	}
+	var res domain.Article
+	err = json.Unmarshal(val, &res)
+	return res, err
+}
+
+func (a *ArticleRedisCache) SetPub(ctx context.Context, art domain.Article) error {
+	val, err := json.Marshal(art)
+	if err != nil {
+		return err
+	}
+	return a.client.Set(ctx, a.pubKey(art.Id), val, time.Minute*10).Err()
+}
+
+func NewArticleRedisCache(client redis.Cmdable) ArticleCache {
+	return &ArticleRedisCache{
 		client: client,
 	}
 }
 
-func (r *RedisArticleCache) GetPub(ctx context.Context, id int64) (domain.Article, error) {
-	// 可以直接使用 Bytes 方法来获得 []byte
-	data, err := r.client.Get(ctx, r.readerArtKey(id)).Bytes()
-	if err != nil {
-		return domain.Article{}, err
-	}
-	var res domain.Article
-	err = json.Unmarshal(data, &res)
-	return res, err
-}
-
-func (r *RedisArticleCache) SetPub(ctx context.Context, art domain.Article) error {
-	data, err := json.Marshal(art)
-	if err != nil {
-		return err
-	}
-	return r.client.Set(ctx, r.readerArtKey(art.Id),
-		data,
-		// 设置长过期时间
-		time.Minute*30).Err()
-}
-
-func (r *RedisArticleCache) Get(ctx context.Context, id int64) (domain.Article, error) {
-	// 可以直接使用 Bytes 方法来获得 []byte
-	data, err := r.client.Get(ctx, r.authorArtKey(id)).Bytes()
-	if err != nil {
-		return domain.Article{}, err
-	}
-	var res domain.Article
-	err = json.Unmarshal(data, &res)
-	return res, err
-}
-
-func (r *RedisArticleCache) Set(ctx context.Context, art domain.Article) error {
-	data, err := json.Marshal(art)
-	if err != nil {
-		return err
-	}
-	return r.client.Set(ctx, r.authorArtKey(art.Id), data, time.Minute).Err()
-}
-
-func (r *RedisArticleCache) DelFirstPage(ctx context.Context, author int64) error {
-	return r.client.Del(ctx, r.firstPageKey(author)).Err()
-}
-
-func (r *RedisArticleCache) GetFirstPage(ctx context.Context, author int64) ([]domain.Article, error) {
-	bs, err := r.client.Get(ctx, r.firstPageKey(author)).Bytes()
+func (a *ArticleRedisCache) GetFirstPage(ctx context.Context, uid int64) ([]domain.Article, error) {
+	key := a.firstKey(uid)
+	//val, err := a.client.Get(ctx, firstKey).Result()
+	val, err := a.client.Get(ctx, key).Bytes()
 	if err != nil {
 		return nil, err
 	}
-	var arts []domain.Article
-	err = json.Unmarshal(bs, &arts)
-	return arts, err
+	var res []domain.Article
+	err = json.Unmarshal(val, &res)
+	return res, err
 }
 
-func (r *RedisArticleCache) SetFirstPage(ctx context.Context, author int64, arts []domain.Article) error {
-	for i := range arts {
-		// 只缓存摘要部分
+func (a *ArticleRedisCache) SetFirstPage(ctx context.Context, uid int64, arts []domain.Article) error {
+	for i := 0; i < len(arts); i++ {
 		arts[i].Content = arts[i].Abstract()
 	}
-	bs, err := json.Marshal(arts)
+	key := a.firstKey(uid)
+	val, err := json.Marshal(arts)
 	if err != nil {
 		return err
 	}
-	return r.client.Set(ctx, r.firstPageKey(author),
-		bs, time.Minute*10).Err()
+	return a.client.Set(ctx, key, val, time.Minute*10).Err()
 }
 
-// 创作端的缓存设置
-func (r *RedisArticleCache) authorArtKey(id int64) string {
-	return fmt.Sprintf("article:author:%d", id)
+func (a *ArticleRedisCache) pubKey(id int64) string {
+	return fmt.Sprintf("article:pub:detail:%d", id)
 }
 
-// 读者端的缓存设置
-func (r *RedisArticleCache) readerArtKey(id int64) string {
-	return fmt.Sprintf("article:reader:%d", id)
+func (a *ArticleRedisCache) key(id int64) string {
+	return fmt.Sprintf("article:detail:%d", id)
 }
 
-func (r *RedisArticleCache) firstPageKey(author int64) string {
-	return fmt.Sprintf("article:first_page:%d", author)
+func (a *ArticleRedisCache) firstKey(uid int64) string {
+	return fmt.Sprintf("article:first_page:%d", uid)
 }
