@@ -1,16 +1,50 @@
 package dao
 
 import (
-	"basic-go/webook/internal/repository/dao/article"
+	"context"
+	"time"
 
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 	"gorm.io/gorm"
 )
 
-func InitTable(db *gorm.DB) error {
-	return db.AutoMigrate(
-		&User{},
-		&article.Article{},
-		&article.PublishedArticle{},
-		//&Job{},
+func InitTables(db *gorm.DB) error {
+	// 严格来说，这个不是优秀实践
+	return db.AutoMigrate(&User{},
+		&Article{},
+		&PublishedArticle{},
+		&AsyncSms{},
+		&Job{},
 	)
+}
+
+func InitCollection(mdb *mongo.Database) error {
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+	col := mdb.Collection("articles")
+	_, err := col.Indexes().CreateMany(ctx, []mongo.IndexModel{
+		{
+			Keys:    bson.D{bson.E{"id", 1}},
+			Options: options.Index().SetUnique(true),
+		},
+		{
+			Keys: bson.D{bson.E{"author_id", 1}},
+		},
+	})
+	if err != nil {
+		return err
+	}
+	liveCol := mdb.Collection("published_articles")
+	_, err = liveCol.Indexes().CreateMany(ctx, []mongo.IndexModel{
+		{
+			Keys:    bson.D{bson.E{"id", 1}},
+			Options: options.Index().SetUnique(true),
+		},
+		{
+			Keys: bson.D{bson.E{"author_id", 1}},
+		},
+	})
+	return err
 }
