@@ -1,9 +1,3 @@
-// Package dao -----------------------------
-// @file      : user.go
-// @author    : hcjjj
-// @contact   : hcjjj@foxmail.com
-// @time      : 2024-02-25 19:39
-// -------------------------------------------
 package dao
 
 import (
@@ -13,25 +7,32 @@ import (
 	"time"
 
 	"github.com/go-sql-driver/mysql"
-
 	"gorm.io/gorm"
 )
 
 var (
-	ErrUserDuplicate = errors.New("邮箱或手机号码冲突")
-	ErrUserNotFound  = gorm.ErrRecordNotFound
+	ErrDuplicateEmail = errors.New("邮箱冲突")
+	ErrRecordNotFound = gorm.ErrRecordNotFound
 )
 
+//go:generate mockgen -source=./user.go -package=daomocks -destination=./mocks/user.mock.go UserDAO
 type UserDAO interface {
-	FindByPhone(ctx context.Context, phone string) (User, error)
-	FindByEmail(ctx context.Context, email string) (User, error)
-	FindById(ctx context.Context, id int64) (User, error)
 	Insert(ctx context.Context, u User) error
-	FindByWechat(ctx context.Context, openID string) (User, error)
+	FindByEmail(ctx context.Context, email string) (User, error)
+	UpdateById(ctx context.Context, entity User) error
+	FindById(ctx context.Context, uid int64) (User, error)
+	FindByPhone(ctx context.Context, phone string) (User, error)
+	FindByWechat(ctx context.Context, openId string) (User, error)
 }
 
 type GORMUserDAO struct {
 	db *gorm.DB
+}
+
+func (dao *GORMUserDAO) FindByWechat(ctx context.Context, openId string) (User, error) {
+	var u User
+	err := dao.db.WithContext(ctx).Where("wechat_open_id=?", openId).First(&u).Error
+	return u, err
 }
 
 func NewUserDAO(db *gorm.DB) UserDAO {
@@ -40,69 +41,84 @@ func NewUserDAO(db *gorm.DB) UserDAO {
 	}
 }
 
-func (dao *GORMUserDAO) FindByWechat(ctx context.Context, openID string) (User, error) {
-	var u User
-	err := dao.db.WithContext(ctx).Where("wechat_open_id = ?", openID).First(&u).Error
-	//err := dao.db.WithContext(ctx).First(&u, "email = ?", email).Error
-	return u, err
-}
-
-func (dao *GORMUserDAO) FindByPhone(ctx context.Context, phone string) (User, error) {
-	var u User
-	err := dao.db.WithContext(ctx).Where("phone = ?", phone).First(&u).Error
-	return u, err
+func (dao *GORMUserDAO) Insert(ctx context.Context, u User) error {
+	now := time.Now().UnixMilli()
+	u.Ctime = now
+	u.Utime = now
+	err := dao.db.WithContext(ctx).Create(&u).Error
+	if me, ok := err.(*mysql.MySQLError); ok {
+		const duplicateErr uint16 = 1062
+		if me.Number == duplicateErr {
+			// 用户冲突，邮箱冲突
+			return ErrDuplicateEmail
+		}
+	}
+	return err
 }
 
 func (dao *GORMUserDAO) FindByEmail(ctx context.Context, email string) (User, error) {
 	var u User
-	err := dao.db.WithContext(ctx).Where("email = ?", email).First(&u).Error
+	err := dao.db.WithContext(ctx).Where("email=?", email).First(&u).Error
 	return u, err
 }
 
-func (dao *GORMUserDAO) FindById(ctx context.Context, id int64) (User, error) {
-	var u User
-	err := dao.db.WithContext(ctx).Where("`id` = ?", id).First(&u).Error
-	return u, err
+func (dao *GORMUserDAO) UpdateById(ctx context.Context, entity User) error {
+
+	// 这种写法依赖于 GORM 的零值和主键更新特性
+	// Update 非零值 WHERE id = ?
+	//return dao.db.WithContext(ctx).Updates(&entity).Error
+	return dao.db.WithContext(ctx).Model(&entity).Where("id = ?", entity.Id).
+		Updates(map[string]any{
+			"utime":    time.Now().UnixMilli(),
+			"nickname": entity.Nickname,
+			"birthday": entity.Birthday,
+			"about_me": entity.AboutMe,
+		}).Error
 }
 
-func (dao *GORMUserDAO) Insert(ctx context.Context, u User) error {
-	// 存毫秒数
-	now := time.Now().UnixMilli()
-	u.Utime = now
-	u.Ctime = now
-	err := dao.db.WithContext(ctx).Create(&u).Error
-	// 判断是 mysql 的错误 和底层强耦合的代码
-	if mysqlErr, ok := err.(*mysql.MySQLError); ok {
-		// MySQL唯一键冲突错误码
-		const uniqueConflictsErrNo uint16 = 1062
-		if mysqlErr.Number == uniqueConflictsErrNo {
-			// 邮箱冲突 or 手机号码冲突
-			return ErrUserDuplicate
-		}
-	}
-	// 为什么不先查询来判断邮箱是否已存在？
-	// 并发查询的时候多个发现不存在，后插入的出现错误❌
-	// 加锁的话会有性能问题，邮箱冲突的情况很少
-	return err
+func (dao *GORMUserDAO) FindById(ctx context.Context, uid int64) (User, error) {
+	var res User
+	err := dao.db.WithContext(ctx).Where("id = ?", uid).First(&res).Error
+	return res, err
 }
 
-// User 直接对于数据库表结构，两者一一对应
-// 如 entity modle PO (persistent object) ...
+func (dao *GORMUserDAO) FindByPhone(ctx context.Context, phone string) (User, error) {
+	var res User
+	err := dao.db.WithContext(ctx).Where("phone = ?", phone).First(&res).Error
+	return res, err
+}
+
 type User struct {
 	Id int64 `gorm:"primaryKey,autoIncrement"`
-	// 唯一索引 允许有多个空值
-	// 但是不能有多个 ""
+	// 代表这是一个可以为 NULL 的列
+	//Email    *string
 	Email    sql.NullString `gorm:"unique"`
 	Password string
-	Phone    sql.NullString `gorm:"unique"`
-	Birthday sql.NullInt64
-	Nickname sql.NullString
-	AboutMe  sql.NullString `gorm:"type:varchar(1024)"`
-	// 创建时间和更新时间，毫秒数，UTC
-	// 避免应用代码和数据库时区的不一致性
+
+	Nickname string `gorm:"type=varchar(128)"`
+	// YYYY-MM-DD
+	Birthday int64
+	AboutMe  string `gorm:"type=varchar(4096)"`
+
+	// 代表这是一个可以为 NULL 的列
+	Phone sql.NullString `gorm:"unique"`
+
+	// 1 如果查询要求同时使用 openid 和 unionid，就要创建联合唯一索引
+	// 2 如果查询只用 openid，那么就在 openid 上创建唯一索引，或者 <openid, unionId> 联合索引
+	// 3 如果查询只用 unionid，那么就在 unionid 上创建唯一索引，或者 <unionid, openid> 联合索引
+	WechatOpenId  sql.NullString `gorm:"unique"`
+	WechatUnionId sql.NullString
+
+	// 时区，UTC 0 的毫秒数
+	// 创建时间
 	Ctime int64
+	// 更新时间
 	Utime int64
-	// 微信的字段
-	WechatUnionID sql.NullString
-	WechatOpenID  sql.NullString `gorm:"unique"`
+
+	// json 存储
+	//Addr string
 }
+
+//type Address struct {
+//	Uid
+//}
