@@ -1,11 +1,10 @@
 package web
 
 import (
-	domain2 "basic-go/webook/interactive/domain"
-	service2 "basic-go/webook/interactive/service"
+	intrv1 "basic-go/webook/api/proto/gen/intr/v1"
 	"basic-go/webook/internal/domain"
 	"basic-go/webook/internal/service"
-	ijwt "basic-go/webook/internal/web/jwt"
+	"basic-go/webook/internal/web/jwt"
 	"basic-go/webook/pkg/ginx"
 	"basic-go/webook/pkg/logger"
 	"fmt"
@@ -17,344 +16,317 @@ import (
 	"time"
 )
 
-var _ handler = (*ArticleHandler)(nil)
-
 type ArticleHandler struct {
 	svc     service.ArticleService
+	intrSvc intrv1.InteractiveServiceClient
 	l       logger.LoggerV1
-	intrSvc service2.InteractiveService
 	biz     string
 }
 
-func NewArticleHandler(svc service.ArticleService,
-	l logger.LoggerV1) *ArticleHandler {
+func NewArticleHandler(l logger.LoggerV1,
+	svc service.ArticleService,
+	intrSvc intrv1.InteractiveServiceClient) *ArticleHandler {
 	return &ArticleHandler{
-		svc: svc,
-		l:   l,
-		biz: "article",
+		l:       l,
+		svc:     svc,
+		intrSvc: intrSvc,
+		biz:     "article",
 	}
 }
 
 func (h *ArticleHandler) RegisterRoutes(server *gin.Engine) {
 	g := server.Group("/articles")
-	// 修改
-	//g.PUT("/")
-	// 新增
-	//g.POST("/")
-	// g.DELETE("/a_id")
 
-	g.POST("/edit", h.Edit)
-	g.POST("/withdraw", h.Withdraw)
-	g.POST("/publish", h.Publish)
-	// 创作者的查询接口
-	// 这个是获取数据的接口，理论上来说（遵循 RESTful 规范），应该是用 GET 方法
-	// GET localhost/articles => List 接口
-	g.POST("/list",
-		ginx.WrapBodyAndToken[ListReq, ijwt.UserClaims](h.List))
-	g.GET("/detail/:id", ginx.WrapToken[ijwt.UserClaims](h.Detail))
+	//g.PUT("/", h.Edit)
+	g.POST("/edit", ginx.WrapBodyAndClaims(h.Edit))
+	g.POST("/publish", ginx.WrapBodyAndClaims(h.Publish))
+	g.POST("/withdraw", ginx.WrapBodyAndClaims(h.Withdraw))
+
+	// 创作者接口
+	g.GET("/detail/:id", h.Detail)
+	// 按照道理来说，这边就是 GET 方法
+	// /list?offset=?&limit=?
+	g.POST("/list", h.List)
 
 	pub := g.Group("/pub")
-	pub.GET("/:id", h.PubDetail, func(ctx *gin.Context) {
-		// 增加阅读计数。
-		//go func() {
-		//	// 开一个 goroutine，异步去执行
-		//	er := a.intrSvc.IncrReadCnt(ctx, a.biz, art.Id)
-		//	if er != nil {
-		//		a.l.Error("增加阅读计数失败",
-		//			logger.Int64("aid", art.Id),
-		//			logger.Error(err))
-		//	}
-		//}()
-	})
-	// 点赞是这个接口，取消点赞也是这个接口
-	// RESTful 风格
-	//pub.POST("/like/:id", ginx.WrapBodyAndToken[LikeReq,
-	//	ijwt.UserClaims](h.Like))
-	pub.POST("/like", ginx.WrapBodyAndToken[LikeReq,
-		ijwt.UserClaims](h.Like))
-	//pub.POST("/cancel_like", ginx.WrapBodyAndToken[LikeReq,
-	//	ijwt.UserClaims](h.Like))
+	pub.GET("/:id", h.PubDetail)
+	// 传入一个参数，true 就是点赞, false 就是不点赞
+	pub.POST("/like", ginx.WrapBodyAndClaims(h.Like))
+	pub.POST("/collect", ginx.WrapBodyAndClaims(h.Collect))
 }
 
-func (a *ArticleHandler) Like(ctx *gin.Context, req LikeReq, uc ijwt.UserClaims) (ginx.Result, error) {
-	var err error
-	if req.Like {
-		err = a.intrSvc.Like(ctx, a.biz, req.Id, uc.Id)
-	} else {
-		err = a.intrSvc.CancelLike(ctx, a.biz, req.Id, uc.Id)
-	}
-
-	if err != nil {
-		return ginx.Result{
-			Code: 5,
-			Msg:  "系统错误",
-		}, err
-	}
-	return ginx.Result{Msg: "OK"}, nil
-}
-
-func (a *ArticleHandler) PubDetail(ctx *gin.Context) {
-	idstr := ctx.Param("id")
-	id, err := strconv.ParseInt(idstr, 10, 64)
-	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 4,
-			Msg:  "参数错误",
-		})
-		a.l.Error("前端输入的 ID 不对", logger.Error(err))
-		return
-	}
-
-	uc := ctx.MustGet("users").(ijwt.UserClaims)
-	var eg errgroup.Group
-	var art domain.Article
-	eg.Go(func() error {
-
-		art, err = a.svc.GetPublishedById(ctx, id, uc.Id)
-		return err
-	})
-
-	var intr domain2.Interactive
-	eg.Go(func() error {
-		// 这个地方可以容忍错误
-		intr, err = a.intrSvc.Get(ctx, a.biz, id, uc.Id)
-		// 这种是容错的写法
-		//if err != nil {
-		//	// 记录日志
-		//}
-		//return nil
-		return err
-	})
-
-	// 在这儿等，要保证前面两个
-	err = eg.Wait()
-	if err != nil {
-		// 代表查询出错了
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "系统错误",
-		})
-		return
-	}
-
-	// 增加阅读计数。
-	go func() {
-		// 你都异步了，怎么还说有巨大的压力呢？
-		// 开一个 goroutine，异步去执行
-		er := a.intrSvc.IncrReadCnt(ctx, a.biz, art.Id)
-		if er != nil {
-			a.l.Error("增加阅读计数失败",
-				logger.Int64("aid", art.Id),
-				logger.Error(err))
-		}
-	}()
-
-	// ctx.Set("art", art)
-
-	// 这个功能是不是可以让前端，主动发一个 HTTP 请求，来增加一个计数？
-	ctx.JSON(http.StatusOK, Result{
-		Data: ArticleVO{
-			Id:      art.Id,
-			Title:   art.Title,
-			Status:  art.Status.ToUint8(),
-			Content: art.Content,
-			// 要把作者信息带出去
-			Author:     art.Author.Name,
-			Ctime:      art.Ctime.Format(time.DateTime),
-			Utime:      art.Utime.Format(time.DateTime),
-			Liked:      intr.Liked,
-			Collected:  intr.Collected,
-			LikeCnt:    intr.LikeCnt,
-			ReadCnt:    intr.ReadCnt,
-			CollectCnt: intr.CollectCnt,
+// Edit 接收 Article 输入，返回一个 ID，文章的 ID
+func (h *ArticleHandler) Edit(ctx *gin.Context,
+	req ArticleEditReq, uc jwt.UserClaims) (ginx.Result, error) {
+	id, err := h.svc.Save(ctx, domain.Article{
+		Id:      req.Id,
+		Title:   req.Title,
+		Content: req.Content,
+		Author: domain.Author{
+			Id: uc.Uid,
 		},
 	})
-}
-
-func (a *ArticleHandler) Detail(ctx *gin.Context, usr ijwt.UserClaims) (ginx.Result, error) {
-	idstr := ctx.Param("id")
-	id, err := strconv.ParseInt(idstr, 10, 64)
 	if err != nil {
-		//ctx.JSON(http.StatusOK, )
-		//a.l.Error("前端输入的 ID 不对", logger.Error(err))
 		return ginx.Result{
-			Code: 4,
-			Msg:  "参数错误",
+			Msg: "系统错误",
 		}, err
-	}
-	art, err := a.svc.GetById(ctx, id)
-	if err != nil {
-		//ctx.JSON(http.StatusOK, )
-		//a.l.Error("获得文章信息失败", logger.Error(err))
-		return ginx.Result{
-			Code: 5,
-			Msg:  "系统错误",
-		}, err
-	}
-	// 这是不借助数据库查询来判定的方法
-	if art.Author.Id != usr.Id {
-		//ctx.JSON(http.StatusOK)
-		// 如果公司有风控系统，这个时候就要上报这种非法访问的用户了。
-		//a.l.Error("非法访问文章，创作者 ID 不匹配",
-		//	logger.Int64("uid", usr.Id))
-		return ginx.Result{
-			Code: 4,
-			// 也不需要告诉前端究竟发生了什么
-			Msg: "输入有误",
-		}, fmt.Errorf("非法访问文章，创作者 ID 不匹配 %d", usr.Id)
 	}
 	return ginx.Result{
-		Data: ArticleVO{
-			Id:    art.Id,
-			Title: art.Title,
-			// 不需要这个摘要信息
-			//Abstract: art.Abstract(),
-			Status:  art.Status.ToUint8(),
-			Content: art.Content,
-			// 这个是创作者看自己的文章列表，也不需要这个字段
-			//Author: art.Author
-			Ctime: art.Ctime.Format(time.DateTime),
-			Utime: art.Utime.Format(time.DateTime),
-		},
+		Data: id,
 	}, nil
 }
 
-func (h *ArticleHandler) Publish(ctx *gin.Context) {
-	var req ArticleReq
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	c := ctx.MustGet("claims")
-	claims, ok := c.(*ijwt.UserClaims)
-	if !ok {
-		// 你可以考虑监控住这里
-		//ctx.AbortWithStatus(http.StatusUnauthorized)
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "系统错误",
-		})
-		h.l.Error("未发现用户的 session 信息")
-		return
-	}
-
-	id, err := h.svc.Publish(ctx, req.toDomain(claims.Id))
-	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "系统错误",
-		})
-		// 打日志？
-		h.l.Error("发表帖子失败", logger.Error(err))
-		return
-	}
-	ctx.JSON(http.StatusOK, Result{
-		Data: id,
-	})
-}
-
-func (h *ArticleHandler) Withdraw(ctx *gin.Context) {
-	type Req struct {
-		Id int64
-	}
-	var req Req
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	c := ctx.MustGet("claims")
-	claims, ok := c.(*ijwt.UserClaims)
-	if !ok {
-		// 你可以考虑监控住这里
-		//ctx.AbortWithStatus(http.StatusUnauthorized)
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "系统错误",
-		})
-		h.l.Error("未发现用户的 session 信息")
-		return
-	}
-
-	// 检测输入，跳过这一步
-	// 调用 svc 的代码
-	err := h.svc.Withdraw(ctx, domain.Article{
-		Id: req.Id,
+func (h *ArticleHandler) Publish(ctx *gin.Context,
+	req PublishReq,
+	uc jwt.UserClaims) (ginx.Result, error) {
+	//val, ok := ctx.Get("user")
+	//if !ok {
+	//	ctx.JSON(http.StatusOK, Result{
+	//		Code: 4,
+	//		Msg:  "未登录",
+	//	})
+	//	return
+	//}
+	id, err := h.svc.Publish(ctx, domain.Article{
+		Id:      req.Id,
+		Title:   req.Title,
+		Content: req.Content,
 		Author: domain.Author{
-			Id: claims.Id,
+			Id: uc.Uid,
 		},
 	})
 	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
+		return ginx.Result{
 			Msg:  "系统错误",
-		})
-		// 打日志？
-		h.l.Error("保存帖子失败", logger.Error(err))
-		return
-	}
-	ctx.JSON(http.StatusOK, Result{
-		Msg: "OK",
-	})
-}
-
-func (h *ArticleHandler) Edit(ctx *gin.Context) {
-	var req ArticleReq
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	c := ctx.MustGet("claims")
-	claims, ok := c.(*ijwt.UserClaims)
-	if !ok {
-		// 你可以考虑监控住这里
-		//ctx.AbortWithStatus(http.StatusUnauthorized)
-		ctx.JSON(http.StatusOK, Result{
 			Code: 5,
-			Msg:  "系统错误",
-		})
-		h.l.Error("未发现用户的 session 信息")
-		return
+		}, fmt.Errorf("发表文章失败 aid %d, uid %d %w", uc.Uid, req.Id, err)
 	}
-	// 检测输入，跳过这一步
-	// 调用 svc 的代码
-	id, err := h.svc.Save(ctx, req.toDomain(claims.Id))
-	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "系统错误",
-		})
-		// 打日志？
-		h.l.Error("保存帖子失败", logger.Error(err))
-		return
-	}
-	ctx.JSON(http.StatusOK, Result{
+	return ginx.Result{
 		Data: id,
-	})
+	}, nil
 }
 
-func (h *ArticleHandler) List(ctx *gin.Context, req ListReq, uc ijwt.UserClaims) (ginx.Result, error) {
-	res, err := h.svc.List(ctx, uc.Id, req.Offset, req.Limit)
+func (h *ArticleHandler) Withdraw(ctx *gin.Context,
+	req ArticleWithdrawReq, uc jwt.UserClaims) (ginx.Result, error) {
+	err := h.svc.Withdraw(ctx, uc.Uid, req.Id)
 	if err != nil {
 		return ginx.Result{
+			Msg:  "系统错误",
+			Code: 5,
+		}, err
+	}
+	return ginx.Result{
+		Msg: "OK",
+	}, nil
+}
+
+func (h *ArticleHandler) List(ctx *gin.Context) {
+	var page Page
+	if err := ctx.Bind(&page); err != nil {
+		return
+	}
+	// 我要不要检测一下？
+	uc := ctx.MustGet("user").(jwt.UserClaims)
+	arts, err := h.svc.GetByAuthor(ctx, uc.Uid, page.Offset, page.Limit)
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{
 			Code: 5,
 			Msg:  "系统错误",
-		}, nil
+		})
+		h.l.Error("查找文章列表失败",
+			logger.Error(err),
+			logger.Int("offset", page.Offset),
+			logger.Int("limit", page.Limit),
+			logger.Int64("uid", uc.Uid))
+		return
 	}
-	// 在列表页，不显示全文，只显示一个"摘要"
-	// 比如说，简单的摘要就是前几句话
-	// 强大的摘要是 AI 帮你生成的
+	ctx.JSON(http.StatusOK, ginx.Result{
+		Data: slice.Map[domain.Article, ArticleVo](arts, func(idx int, src domain.Article) ArticleVo {
+			return ArticleVo{
+				Id:       src.Id,
+				Title:    src.Title,
+				Abstract: src.Abstract(),
+
+				//Content:  src.Content,
+				AuthorId: src.Author.Id,
+				// 列表，你不需要
+				Status: src.Status.ToUint8(),
+				Ctime:  src.Ctime.Format(time.DateTime),
+				Utime:  src.Utime.Format(time.DateTime),
+			}
+		}),
+	})
+}
+
+func (h *ArticleHandler) Detail(ctx *gin.Context) {
+	idstr := ctx.Param("id")
+	id, err := strconv.ParseInt(idstr, 10, 64)
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{
+			Msg:  "id 参数错误",
+			Code: 4,
+		})
+		h.l.Warn("查询文章失败，id 格式不对",
+			logger.String("id", idstr),
+			logger.Error(err))
+		return
+	}
+	art, err := h.svc.GetById(ctx, id)
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{
+			Msg:  "系统错误",
+			Code: 5,
+		})
+		h.l.Error("查询文章失败",
+			logger.Int64("id", id),
+			logger.Error(err))
+		return
+	}
+	uc := ctx.MustGet("user").(jwt.UserClaims)
+	if art.Author.Id != uc.Uid {
+		// 有人在搞鬼
+		ctx.JSON(http.StatusOK, ginx.Result{
+			Msg:  "系统错误",
+			Code: 5,
+		})
+		h.l.Error("非法查询文章",
+			logger.Int64("id", id),
+			logger.Int64("uid", uc.Uid))
+		return
+	}
+
+	vo := ArticleVo{
+		Id:    art.Id,
+		Title: art.Title,
+		//Abstract: art.Abstract(),
+
+		Content:  art.Content,
+		AuthorId: art.Author.Id,
+		// 列表，你不需要
+		Status: art.Status.ToUint8(),
+		Ctime:  art.Ctime.Format(time.DateTime),
+		Utime:  art.Utime.Format(time.DateTime),
+	}
+	ctx.JSON(http.StatusOK, ginx.Result{Data: vo})
+}
+
+func (h *ArticleHandler) PubDetail(ctx *gin.Context) {
+	idstr := ctx.Param("id")
+	id, err := strconv.ParseInt(idstr, 10, 64)
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{
+			Msg:  "id 参数错误",
+			Code: 4,
+		})
+		h.l.Warn("查询文章失败，id 格式不对",
+			logger.String("id", idstr),
+			logger.Error(err))
+		return
+	}
+
+	var (
+		eg   errgroup.Group
+		art  domain.Article
+		intr *intrv1.GetResponse
+	)
+
+	uc := ctx.MustGet("user").(jwt.UserClaims)
+	eg.Go(func() error {
+		var er error
+		art, er = h.svc.GetPubById(ctx, id, uc.Uid)
+		return er
+	})
+	eg.Go(func() error {
+		var er error
+		intr, er = h.intrSvc.Get(ctx, &intrv1.GetRequest{
+			Biz: h.biz, BizId: id, Uid: uc.Uid,
+		})
+		return er
+	})
+
+	// 等待结果
+	err = eg.Wait()
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{
+			Msg:  "系统错误",
+			Code: 5,
+		})
+		h.l.Error("查询文章失败，系统错误",
+			logger.Int64("aid", id),
+			logger.Int64("uid", uc.Uid),
+			logger.Error(err))
+		return
+	}
+
+	//go func() {
+	// 1. 如果你想摆脱原本主链路的超时控制，你就创建一个新的
+	// 2. 如果你不想，你就用 ctx
+	//newCtx, cancel := context.WithTimeout(context.Background(), time.Second)
+	//defer cancel()
+	//er := h.intrSvc.IncrReadCnt(newCtx, h.biz, art.Id)
+	//if er != nil {
+	//	h.l.Error("更新阅读数失败",
+	//		logger.Int64("aid", art.Id),
+	//		logger.Error(err))
+	//}
+	//}()
+
+	ctx.JSON(http.StatusOK, ginx.Result{
+		Data: ArticleVo{
+			Id:    art.Id,
+			Title: art.Title,
+
+			Content:    art.Content,
+			AuthorId:   art.Author.Id,
+			AuthorName: art.Author.Name,
+			ReadCnt:    intr.Intr.ReadCnt,
+			CollectCnt: intr.Intr.CollectCnt,
+			LikeCnt:    intr.Intr.LikeCnt,
+			Liked:      intr.Intr.Liked,
+			Collected:  intr.Intr.Collected,
+
+			Status: art.Status.ToUint8(),
+			Ctime:  art.Ctime.Format(time.DateTime),
+			Utime:  art.Utime.Format(time.DateTime),
+		},
+	})
+}
+
+func (h *ArticleHandler) Like(c *gin.Context,
+	req ArticleLikeReq, uc jwt.UserClaims) (ginx.Result, error) {
+	var err error
+	if req.Like {
+		// 点赞
+		_, err = h.intrSvc.Like(c, &intrv1.LikeRequest{
+			Biz: h.biz, BizId: req.Id, Uid: uc.Uid,
+		})
+	} else {
+		// 取消点赞
+		_, err = h.intrSvc.CancelLike(c, &intrv1.CancelLikeRequest{
+			Biz: h.biz, BizId: req.Id, Uid: uc.Uid,
+		})
+	}
+	if err != nil {
+		return ginx.Result{
+			Code: 5, Msg: "系统错误",
+		}, err
+	}
 	return ginx.Result{
-		Data: slice.Map[domain.Article, ArticleVO](res,
-			func(idx int, src domain.Article) ArticleVO {
-				return ArticleVO{
-					Id:       src.Id,
-					Title:    src.Title,
-					Abstract: src.Abstract(),
-					Status:   src.Status.ToUint8(),
-					// 这个列表请求，不需要返回内容
-					//Content: src.Content,
-					// 这个是创作者看自己的文章列表，也不需要这个字段
-					//Author: src.Author
-					Ctime: src.Ctime.Format(time.DateTime),
-					Utime: src.Utime.Format(time.DateTime),
-				}
-			}),
+		Msg: "OK",
+	}, nil
+}
+
+func (h *ArticleHandler) Collect(ctx *gin.Context,
+	req ArticleCollectReq, uc jwt.UserClaims) (ginx.Result, error) {
+	_, err := h.intrSvc.Collect(ctx, &intrv1.CollectRequest{
+		Biz: h.biz, BizId: req.Id, Uid: uc.Uid, Cid: req.Cid,
+	})
+	if err != nil {
+		return ginx.Result{
+			Code: 5, Msg: "系统错误",
+		}, err
+	}
+	return ginx.Result{
+		Msg: "OK",
 	}, nil
 }
