@@ -1,59 +1,46 @@
 package web
 
-import "basic-go/webook/internal/domain"
+type ArticleVo struct {
+	Id         int64  `json:"id,omitempty"`
+	Title      string `json:"title,omitempty"`
+	Abstract   string `json:"abstract,omitempty"`
+	Content    string `json:"content,omitempty"`
+	AuthorId   int64  `json:"authorId,omitempty"`
+	AuthorName string `json:"authorName,omitempty"`
+	Status     uint8  `json:"status,omitempty"`
+	Ctime      string `json:"ctime,omitempty"`
+	Utime      string `json:"utime,omitempty"`
 
-// VO view object，就是对标前端的
-
-type LikeReq struct {
-	Id int64 `json:"id"`
-	// 点赞和取消点赞，准备复用这个
-	Like bool `json:"like"`
+	ReadCnt    int64 `json:"readCnt"`
+	LikeCnt    int64 `json:"likeCnt"`
+	CollectCnt int64 `json:"collectCnt"`
+	Liked      bool  `json:"liked"`
+	Collected  bool  `json:"collected"`
 }
 
-type ArticleVO struct {
-	Id    int64  `json:"id"`
-	Title string `json:"title"`
-	// 摘要
-	Abstract string `json:"abstract"`
-	// 内容
-	Content string `json:"content"`
-	// 注意一点，状态这个东西，可以是前端来处理，也可以是后端处理
-	// 0 -> unknown -> 未知状态
-	// 1 -> 未发表，手机 APP 这种涉及到发版的问题，那么后端来处理
-	// 涉及到国际化，也是后端来处理
-	Status uint8  `json:"status"`
-	Author string `json:"author"`
-	// 计数
-	ReadCnt    int64 `json:"read_cnt"`
-	LikeCnt    int64 `json:"like_cnt"`
-	CollectCnt int64 `json:"collect_cnt"`
-
-	// 我个人有没有收藏，有没有点赞
-	Liked     bool `json:"liked"`
-	Collected bool `json:"collected"`
-
-	Ctime string `json:"ctime"`
-	Utime string `json:"utime"`
-}
-
-type ListReq struct {
-	Offset int `json:"offset"`
-	Limit  int `json:"limit"`
-}
-
-type ArticleReq struct {
-	Id      int64  `json:"id"`
+type PublishReq struct {
+	Id      int64
 	Title   string `json:"title"`
 	Content string `json:"content"`
 }
 
-func (req ArticleReq) toDomain(uid int64) domain.Article {
-	return domain.Article{
-		Id:      req.Id,
-		Title:   req.Title,
-		Content: req.Content,
-		Author: domain.Author{
-			Id: uid,
-		},
-	}
+type ArticleEditReq struct {
+	Id      int64
+	Title   string `json:"title"`
+	Content string `json:"content"`
+}
+
+type ArticleWithdrawReq struct {
+	Id int64
+}
+
+type ArticleLikeReq struct {
+	Id int64 `json:"id"`
+	// true 是点赞，false 是不点赞
+	Like bool `json:"like"`
+}
+
+type ArticleCollectReq struct {
+	Id  int64 `json:"id"`
+	Cid int64 `json:"cid"`
 }
