@@ -1,30 +1,11 @@
 package jwt
 
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/golang-jwt/jwt/v5"
-)
+import "github.com/gin-gonic/gin"
 
 type Handler interface {
+	ClearToken(ctx *gin.Context) error
+	ExtractToken(ctx *gin.Context) string
 	SetLoginToken(ctx *gin.Context, uid int64) error
 	SetJWTToken(ctx *gin.Context, uid int64, ssid string) error
-	ClearToken(ctx *gin.Context) error
 	CheckSession(ctx *gin.Context, ssid string) error
-	ExtractToken(ctx *gin.Context) string
-}
-
-type RefreshClaims struct {
-	Uid  int64
-	Ssid string
-	jwt.RegisteredClaims
-	//UserAgent string
-}
-
-type UserClaims struct {
-	jwt.RegisteredClaims
-	// 声明你自己的要放进去 token 里面的数据
-	Id   int64
-	Ssid string
-	// 自己随便加
-	UserAgent string
 }
