@@ -1,110 +1,69 @@
-// Package middleware -----------------------------
-// @file      : login_jwt.go
-// @author    : hcjjj
-// @contact   : hcjjj@foxmail.com
-// @time      : 2024-03-14 11:22
-// -------------------------------------------
 package middleware
 
 import (
+	ijwt "basic-go/webook/internal/web/jwt"
 	"net/http"
 
-	"github.com/golang-jwt/jwt/v5"
-
 	"github.com/gin-gonic/gin"
-
-	ijwt "basic-go/webook/internal/web/jwt"
+	"github.com/golang-jwt/jwt/v5"
 )
 
 type LoginJWTMiddlewareBuilder struct {
-	paths []string
 	ijwt.Handler
 }
 
-func NewLoginJWTMiddlewareBuilder(jwtHdl ijwt.Handler) *LoginJWTMiddlewareBuilder {
+func NewLoginJWTMiddlewareBuilder(hdl ijwt.Handler) *LoginJWTMiddlewareBuilder {
 	return &LoginJWTMiddlewareBuilder{
-		Handler: jwtHdl,
+		Handler: hdl,
 	}
-
 }
 
-// IgnorePaths 中间方法，构建部分
-func (l *LoginJWTMiddlewareBuilder) IgnorePaths(path string) *LoginJWTMiddlewareBuilder {
-	l.paths = append(l.paths, path)
-	return l
-}
-
-// Build 完成构建，返回最终需要的
-func (l *LoginJWTMiddlewareBuilder) Build() gin.HandlerFunc {
-	// 用 Go 的方式编码解码
+func (m *LoginJWTMiddlewareBuilder) CheckLogin() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		// 不需要登录校验的
-		for _, path := range l.paths {
-			if ctx.Request.URL.Path == path {
-				return
-			}
+		path := ctx.Request.URL.Path
+		if path == "/users/signup" ||
+			path == "/users/login" ||
+			path == "/users/login_sms/code/send" ||
+			path == "/users/login_sms" ||
+			path == "/oauth2/wechat/authurl" ||
+			path == "/oauth2/wechat/callback" {
+			// 不需要登录校验
+			return
 		}
-
-		tokenStr := l.ExtractToken(ctx)
-		claims := &ijwt.UserClaims{}
-		// ParseWithClaims 里面，一定要传入指针
-		token, err := jwt.ParseWithClaims(tokenStr, claims, func(token *jwt.Token) (interface{}, error) {
-			return []byte("jaks3jgvkjoiGezwd4QbE9ujPZp0fL8p"), nil
+		tokenStr := m.ExtractToken(ctx)
+		var uc ijwt.UserClaims
+		token, err := jwt.ParseWithClaims(tokenStr, &uc, func(token *jwt.Token) (interface{}, error) {
+			return ijwt.JWTKey, nil
 		})
 		if err != nil {
-			// 没登录
+			// token 不对，token 是伪造的
 			ctx.AbortWithStatus(http.StatusUnauthorized)
 			return
 		}
-		//claims.ExpiresAt.Time.Before(time.Now()) {
-		//	// 过期了
-		//}
-		// err 为 nil，token 不为 nil
-		if token == nil || !token.Valid || claims.Id == 0 {
-			// 没登录
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-			return
-		}
-		if claims.UserAgent != ctx.Request.UserAgent() {
-			// 严重的安全问题
-			// 你是要监控
+		if token == nil || !token.Valid {
+			// 在这里发现 access_token 过期了，生成一个新的 access_token
+
+			// token 解析出来了，但是 token 可能是非法的，或者过期了的
 			ctx.AbortWithStatus(http.StatusUnauthorized)
 			return
 		}
 
-		// 检查 Token 是否在黑名单
-		err = l.CheckSession(ctx, claims.Ssid)
+		// 这里看
+		err = m.CheckSession(ctx, uc.Ssid)
 		if err != nil {
-			// 要么 redis 有问题，要么已经退出登录
+			// token 无效或者 redis 有问题
 			ctx.AbortWithStatus(http.StatusUnauthorized)
 			return
 		}
 
-		// 你以为的退出登录，没有用的
-		//token.Valid = false
-		//// tokenStr 是一个新的字符串
-		//tokenStr, err = token.SignedString([]byte("jaks3jgvkjoiGezwd4QbE9ujPZp0fL8p"))
-		//if err != nil {
-		//	// 记录日志
-		//	log.Println("jwt 续约失败", err)
-		//}
-		//ctx.Header("x-jwt-token", tokenStr)
-
-		// 短的 token 过期了，搞个新的
-		//now := time.Now()
-		// 每十秒钟刷新一次
-		//if claims.ExpiresAt.Sub(now) < time.Second*50 {
-		//	claims.ExpiresAt = jwt.NewNumericDate(time.Now().Add(time.Minute))
-		//	tokenStr, err = token.SignedString([]byte("jaks3jgvkjoiGezwd4QbE9ujPZp0fL8p"))
-		//	if err != nil {
-		//		// 记录日志
-		//		log.Println("jwt 续约失败", err)
-		//	}
-		//	ctx.Header("x-jwt-token", tokenStr)
+		// 可以兼容 Redis 异常的情况
+		// 做好监控，监控有没有 error
+		//if cnt > 0 {
+		//	// token 无效或者 redis 有问题
+		//	ctx.AbortWithStatus(http.StatusUnauthorized)
+		//	return
 		//}
 
-		ctx.Set("claims", claims)
-
-		//ctx.Set("userId", claims.Uid)
+		ctx.Set("user", uc)
 	}
 }
