@@ -1,452 +1,351 @@
-// Package web -----------------------------
-// @file      : user.go
-// @author    : hcjjj
-// @contact   : hcjjj@foxmail.com
-// @time      : 2024-01-29 20:45
-// -------------------------------------------
 package web
 
 import (
+	"net/http"
+	"time"
+
 	"basic-go/webook/internal/domain"
 	"basic-go/webook/internal/errs"
 	"basic-go/webook/internal/service"
 	ijwt "basic-go/webook/internal/web/jwt"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/golang-jwt/jwt/v5"
-
-	"github.com/gin-contrib/sessions"
+	"basic-go/webook/pkg/ginx"
 
 	regexp "github.com/dlclark/regexp2"
+	"github.com/gin-contrib/sessions"
 	"github.com/gin-gonic/gin"
+	"github.com/golang-jwt/jwt/v5"
 )
 
-// UserHandler 定义用户有关的路由
+const (
+	emailRegexPattern = "^\\w+([-+.]\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$"
+	// 和上面比起来，用 ` 看起来就比较清爽
+	passwordRegexPattern = `^(?=.*[A-Za-z])(?=.*\d)(?=.*[$@$!%*#?&])[A-Za-z\d$@$!%*#?&]{8,}$`
+	bizLogin             = "login"
+)
+
 type UserHandler struct {
-	svc         service.UserService
-	codeSvc     service.CodeService
-	emailExp    *regexp.Regexp
-	passwordExp *regexp.Regexp
-	// 组合
 	ijwt.Handler
+	emailRexExp    *regexp.Regexp
+	passwordRexExp *regexp.Regexp
+	svc            service.UserService
+	codeSvc        service.CodeService
 }
 
-func NewUserHandler(svc service.UserService, codeSvc service.CodeService,
-	jwtHdl ijwt.Handler) *UserHandler {
-	const (
-		emailRegexPattern    = "\\w+([-+.]\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*"
-		passwordRegexPattern = "^(?=.*\\d)(?=.*[a-zA-Z])(?=.*[^\\da-zA-Z\\s]).{1,16}$"
-	)
-	// 预编译所需的正则表达式
-	emailExp := regexp.MustCompile(emailRegexPattern, regexp.None)
-	passwordExp := regexp.MustCompile(passwordRegexPattern, regexp.None)
+func NewUserHandler(svc service.UserService,
+	hdl ijwt.Handler,
+	codeSvc service.CodeService) *UserHandler {
 	return &UserHandler{
-		svc:         svc,
-		emailExp:    emailExp,
-		passwordExp: passwordExp,
-		codeSvc:     codeSvc,
-		Handler:     jwtHdl,
+		emailRexExp:    regexp.MustCompile(emailRegexPattern, regexp.None),
+		passwordRexExp: regexp.MustCompile(passwordRegexPattern, regexp.None),
+		svc:            svc,
+		codeSvc:        codeSvc,
+		Handler:        hdl,
 	}
 }
 
-//func (u *UserHandler) RegisterRoutesV1(ug *gin.RouterGroup) {
-//	ug.POST("/signup", u.SignUp)
-//	ug.POST("/login", u.Login)
-//	ug.POST("/edit", u.Edit)
-//	ug.GET("/profile", u.Profile)
-//}
-
-func (u *UserHandler) RegisterRoutes(server *gin.Engine) {
-	//server.POST("/users/signup", u.SignUp)
-	//server.POST("/users/login", u.Login)
-	//server.POST("/users/edit", u.Edit)
-	//server.GET("/users/profile", u.Profile)
+func (h *UserHandler) RegisterRoutes(server *gin.Engine) {
+	// REST 风格
+	//server.POST("/user", h.SignUp)
+	//server.PUT("/user", h.SignUp)
+	//server.GET("/users/:username", h.Profile)
 	ug := server.Group("/users")
-	ug.POST("/signup", u.SignUp)
-	//ug.POST("/login", u.Login)
-	ug.POST("/login", u.LoginJWT)
-	ug.POST("/logout", u.LogoutJWT)
-	//ug.POST("/logout", u.Logout)
-	ug.POST("/edit", u.Edit)
-	//ug.GET("/profile", u.Profile)
-	ug.GET("/profile", u.ProfileJWT)
-	// PUT "/login/sms/code" 发送验证码
-	// POST "/login/sms/code" 校验验证码
-	// POST "/sms/login/code"
-	ug.POST("/login_sms/code/send", u.SendLoginSMSCode)
-	ug.POST("/login_sms", u.LoginSMS)
-	ug.POST("/refresh_token", u.RefreshToken)
+	// POST /users/signup
+	ug.POST("/signup", ginx.WrapBody(h.SignUp))
+	// POST /users/login
+	//ug.POST("/login", h.Login)
+	ug.POST("/login", ginx.WrapBody(h.LoginJWT))
+	ug.POST("/logout", h.LogoutJWT)
+	// POST /users/edit
+	ug.POST("/edit", ginx.WrapBodyAndClaims(h.Edit))
+	// GET /users/profile
+	ug.GET("/profile", ginx.WrapClaims(h.Profile))
+	ug.GET("/refresh_token", h.RefreshToken)
+
+	// 手机验证码登录相关功能
+	ug.POST("/login_sms/code/send", ginx.WrapBody(h.SendSMSLoginCode))
+	ug.POST("/login_sms", ginx.WrapBody(h.LoginSMS))
 }
 
-const biz = "login"
-
-func (u *UserHandler) LoginSMS(ctx *gin.Context) {
-	type Req struct {
-		Phone string `json:"phone"`
-		Code  string `json:"code"`
-	}
-	var req Req
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	// 可以加上各种校验
-
-	ok, err := u.codeSvc.Verify(ctx, biz, req.Phone, req.Code)
+func (h *UserHandler) LoginSMS(ctx *gin.Context,
+	req LoginSMSReq) (ginx.Result, error) {
+	ok, err := h.codeSvc.Verify(ctx, bizLogin, req.Phone, req.Code)
 	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
+		return ginx.Result{
 			Code: 5,
-			Msg:  "系统错误",
-		})
+			Msg:  "系统异常",
+			//Msg: err.Error(),
+		}, err
 	}
 	if !ok {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "验证码有误",
-		})
-		return
+		return ginx.Result{
+			Code: 4,
+			Msg:  "验证码不对，请重新输入",
+		}, nil
 	}
-	// 这个手机号是否会是新用户？
-	user, err := u.svc.FindOrCreate(ctx, req.Phone)
+	u, err := h.svc.FindOrCreate(ctx, req.Phone)
 	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
+		return ginx.Result{
 			Code: 5,
 			Msg:  "系统错误",
-		})
+		}, err
 	}
-	// 登录成功
-	if err = u.SetLoginToken(ctx, user.Id); err != nil {
-		// 记录日志
-		ctx.JSON(http.StatusOK, Result{
+	err = h.SetLoginToken(ctx, u.Id)
+	if err != nil {
+		return ginx.Result{
 			Code: 5,
 			Msg:  "系统错误",
-		})
-		return
+		}, err
 	}
-
-	ctx.JSON(http.StatusOK, Result{
-		Code: 4,
-		Msg:  "登录成功",
-	})
+	return ginx.Result{
+		Msg: "登录成功",
+	}, nil
 }
 
-func (u *UserHandler) SendLoginSMSCode(ctx *gin.Context) {
-	type Req struct {
-		Phone string `json:"phone"`
-	}
-	var req Req
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	// 需要添加使用正则表达式校验 手机号格式
+func (h *UserHandler) SendSMSLoginCode(ctx *gin.Context,
+	req SendSMSCodeReq) (ginx.Result, error) {
+	// 你这边可以校验 Req
 	if req.Phone == "" {
-		ctx.JSON(http.StatusOK, Result{
+		return ginx.Result{
 			Code: 4,
-			Msg:  "输入有误",
-		})
+			Msg:  "请输入手机号码",
+		}, nil
 	}
-	err := u.codeSvc.Send(ctx, biz, req.Phone)
-	// 这边可以细化错误
+	err := h.codeSvc.Send(ctx, bizLogin, req.Phone)
 	switch err {
 	case nil:
-		ctx.JSON(http.StatusOK, Result{
+		return ginx.Result{
 			Msg: "发送成功",
-		})
+		}, nil
 	case service.ErrCodeSendTooMany:
-		ctx.JSON(http.StatusOK, Result{
-			Msg: "发送太频繁，请稍后再试",
-		})
+		// 事实上，防不住有人不知道怎么触发了
+		// 少数这种错误，是可以接受的
+		// 但是频繁出现，就代表有人在搞你的系统
+		return ginx.Result{
+			Code: 4,
+			Msg:  "短信发送太频繁，请稍后再试",
+		}, nil
 	default:
-		ctx.JSON(http.StatusOK, Result{
+		return ginx.Result{
 			Code: 5,
 			Msg:  "系统错误",
-		})
+		}, err
 	}
 }
 
-func (u *UserHandler) SignUp(ctx *gin.Context) {
-	type SignUpReq struct {
-		Email           string `json:"email"`
-		ConfirmPassword string `json:"confirmPassword"`
-		Password        string `json:"password"`
-	}
-	var req SignUpReq
-	// 接收请求
-	// Bind 方法会根据 Content-type 解析数据到 req
-	// 解析出错回直接写回一个 400 错误
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	// 数据校验
-	ok, err := u.emailExp.MatchString(req.Email)
+func (h *UserHandler) SignUp(ctx *gin.Context, req SignUpReq) (ginx.Result, error) {
+	isEmail, err := h.emailRexExp.MatchString(req.Email)
 	if err != nil {
 		ctx.String(http.StatusOK, "系统错误")
-		return
+		return ginx.Result{
+			Code: errs.UserInternalServerError,
+			Msg:  "系统错误",
+		}, err
 	}
-	if !ok {
-		ctx.String(http.StatusOK, "邮箱格式不对")
-		return
+	if !isEmail {
+		return ginx.Result{
+			Code: errs.UserInvalidInput,
+			Msg:  "非法邮箱格式",
+		}, nil
 	}
 
-	if req.ConfirmPassword != req.Password {
-		ctx.String(http.StatusOK, "两次输入的密码不一致")
-		return
+	if req.Password != req.ConfirmPassword {
+		return ginx.Result{
+			Code: errs.UserInvalidInput,
+			Msg:  "两次输入的密码不相等",
+		}, nil
 	}
 
-	ok, err = u.passwordExp.MatchString(req.Password)
+	isPassword, err := h.passwordRexExp.MatchString(req.Password)
 	if err != nil {
-		// 记录日志
-		// ...
-		ctx.String(http.StatusOK, "系统错误")
-		return
+		return ginx.Result{
+			Code: errs.UserInternalServerError,
+			Msg:  "系统错误",
+		}, err
 	}
-	if !ok {
-		ctx.String(http.StatusOK, "至少包含字母、数字、特殊字符，1-16位")
-		return
+	if !isPassword {
+		return ginx.Result{
+			Code: errs.UserInvalidInput,
+			Msg:  "密码必须包含字母、数字、特殊字符",
+		}, nil
 	}
 
-	// 调用一下 svc 的方法
-	err = u.svc.SignUp(ctx, domain.User{
+	err = h.svc.Signup(ctx.Request.Context(), domain.User{
 		Email:    req.Email,
 		Password: req.Password,
 	})
-	if err == service.ErrUserDuplicate {
-		ctx.String(http.StatusOK, "邮箱或者手机号码冲突")
-		return
+	switch err {
+	case nil:
+		return ginx.Result{
+			Msg: "OK",
+		}, nil
+	case service.ErrDuplicateEmail:
+		return ginx.Result{
+			Code: errs.UserDuplicateEmail,
+			Msg:  "邮箱冲突",
+		}, nil
+	default:
+		return ginx.Result{
+			Code: errs.UserInternalServerError,
+			Msg:  "系统错误",
+		}, err
 	}
-	if err != nil {
-		ctx.String(http.StatusOK, "系统异常")
-		return
-	}
-	ctx.String(http.StatusOK, "注册成功")
-	//fmt.Printf("%v\n", req)
 }
 
-// RefreshToken 可以同时刷新长短 token，用 redis 来记录是否有效，即 refresh_token 是一次性的
-// 参考登录校验部分，比较 User-Agent 来增强安全性
-func (u *UserHandler) RefreshToken(ctx *gin.Context) {
-	// 只有这个接口，拿出来的才是 refresh_token，其它地方都是 access token
-	refreshToken := u.ExtractToken(ctx)
-	var rc ijwt.RefreshClaims
-	token, err := jwt.ParseWithClaims(refreshToken, &rc, func(token *jwt.Token) (interface{}, error) {
-		return ijwt.RtKey, nil
-	})
-	if err != nil || !token.Valid {
-		ctx.AbortWithStatus(http.StatusUnauthorized)
-		return
+func (h *UserHandler) LoginJWT(ctx *gin.Context, req LoginJWTReq) (ginx.Result, error) {
+	u, err := h.svc.Login(ctx, req.Email, req.Password)
+	switch err {
+	case nil:
+		err = h.SetLoginToken(ctx, u.Id)
+		if err != nil {
+			return ginx.Result{
+				Code: 5,
+				Msg:  "系统错误",
+			}, err
+		}
+		return ginx.Result{
+			Msg: "OK",
+		}, nil
+	case service.ErrInvalidUserOrPassword:
+		return ginx.Result{Msg: "用户名或者密码错误"}, nil
+	default:
+		return ginx.Result{Msg: "系统错误"}, err
 	}
-	err = u.CheckSession(ctx, rc.Ssid)
-	if err != nil {
-		// 要么 redis 有问题，要么已经退出登录
-		ctx.AbortWithStatus(http.StatusUnauthorized)
-		return
-	}
-	// 搞个新的 access_token
-	err = u.SetJWTToken(ctx, rc.Uid, rc.Ssid)
-	if err != nil {
-		ctx.AbortWithStatus(http.StatusUnauthorized)
-		return
-	}
-	ctx.JSON(http.StatusOK, Result{
-		Msg: "刷新成功",
-	})
 }
 
-func (u *UserHandler) LogoutJWT(ctx *gin.Context) {
-	err := u.ClearToken(ctx)
-	if err != nil {
-		ctx.JSON(http.StatusOK, Result{
-			Code: 5,
-			Msg:  "退出登录失败",
-		})
-		return
-	}
-	ctx.JSON(http.StatusOK, Result{
-		Msg: "退出登录OK",
-	})
-}
+//func (h *UserHandler) Logout(ctx *gin.Context) {
+//	sess := sessions.Default(ctx)
+//	sess.Options(sessions.Options{
+//		MaxAge: -1,
+//	})
+//	sess.Save()
+//}
 
-func (u *UserHandler) LoginJWT(ctx *gin.Context) {
-	type LoginReq struct {
-		Email    string `json:"email"`
-		Password string `json:"password"`
-	}
-	var req LoginReq
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	user, err := u.svc.Login(ctx, req.Email, req.Password)
-	if err == service.ErrInvalidUserOrPassword {
-		ctx.String(http.StatusOK, "邮箱或密码不对")
-		return
-	}
-	if err != nil {
-		ctx.String(http.StatusOK, "系统错误")
-		return
-	}
-	// 到这里就登录成功了
-	// 用 JWT 设置登录态
-	// 生成长短 token
-	if err = u.SetLoginToken(ctx, user.Id); err != nil {
-		ctx.String(http.StatusOK, "系统错误")
-		return
-	}
-	ctx.String(http.StatusOK, "登录成功")
-	return
-}
-
-func (u *UserHandler) Login(ctx *gin.Context) {
-	type LoginReq struct {
-		Email    string `json:"email"`
-		Password string `json:"password"`
-	}
-	var req LoginReq
-	if err := ctx.Bind(&req); err != nil {
-		return
-	}
-	user, err := u.svc.Login(ctx, req.Email, req.Password)
-	//if err == service.ErrInvalidUserOrPassword {
-	//	ctx.String(http.StatusOK, "邮箱或密码不对")
-	//	return
-	//}
-	if err == service.ErrInvalidUserOrPassword {
-		ctx.JSON(http.StatusOK, Result{
-			Code: errs.UserInvalidOrPassword,
-			Msg:  "用户不存在或密码错误",
-		})
-	}
-	if err != nil {
-		ctx.String(http.StatusOK, "系统错误")
-		return
-	}
-
-	// 到这里就登录成功了
-	// 设置 session
-	sess := sessions.Default(ctx)
-	// 设置放在 session 里面的值
-	sess.Set("userId", user.Id)
-	sess.Options(sessions.Options{
-		// 10 min 过期
-		MaxAge: 10 * 60,
-	})
-	sess.Save()
-	ctx.String(http.StatusOK, "登录成功")
-	return
-}
-
-func (u *UserHandler) Edit(ctx *gin.Context) {
+func (h *UserHandler) Login(ctx *gin.Context) {
 	type Req struct {
-		// 注意，其它字段，尤其是密码、邮箱和手机，
-		// 修改都要通过别的手段
-		// 邮箱和手机都要验证
-		// 密码更加不用多说了
-		Nickname string `json:"nickname"`
-		// 2023-01-01
-		Birthday string `json:"birthday"`
-		AboutMe  string `json:"aboutMe"`
+		Email    string `json:"email"`
+		Password string `json:"password"`
 	}
-
 	var req Req
 	if err := ctx.Bind(&req); err != nil {
 		return
 	}
-	// 你可以尝试在这里校验。
-	// 比如说你可以要求 Nickname 必须不为空
-	// 校验规则取决于产品经理
-	if req.Nickname == "" {
-		ctx.JSON(http.StatusOK, Result{Code: 4, Msg: "昵称不能为空"})
-		return
+	u, err := h.svc.Login(ctx, req.Email, req.Password)
+	switch err {
+	case nil:
+		sess := sessions.Default(ctx)
+		sess.Set("userId", u.Id)
+		sess.Options(sessions.Options{
+			// 十分钟
+			MaxAge: 30,
+		})
+		err = sess.Save()
+		if err != nil {
+			ctx.String(http.StatusOK, "系统错误")
+			return
+		}
+		ctx.String(http.StatusOK, "登录成功")
+	case service.ErrInvalidUserOrPassword:
+		ctx.String(http.StatusOK, "用户名或者密码不对")
+	default:
+		ctx.String(http.StatusOK, "系统错误")
 	}
+}
 
-	if len(req.AboutMe) > 1024 {
-		ctx.JSON(http.StatusOK, Result{Code: 4, Msg: "关于我过长"})
-		return
-	}
+func (h *UserHandler) Edit(ctx *gin.Context, req UserEditReq,
+	uc ijwt.UserClaims) (ginx.Result, error) {
+	// 嵌入一段刷新过期时间的代码
+	//sess := sessions.Default(ctx)
+	//sess.Get("uid")
+	// 用户输入不对
 	birthday, err := time.Parse(time.DateOnly, req.Birthday)
 	if err != nil {
-		// 也就是说，其实并没有直接校验具体的格式
-		// 而是如果你能转化过来，那就说明没问题
-		ctx.JSON(http.StatusOK, Result{Code: 4, Msg: "日期格式不对"})
-		return
+		return ginx.Result{
+			Code: 4,
+			Msg:  "生日格式不对",
+		}, err
 	}
-
-	// 这边要处理数据库和缓存了
-	// .....
-	fmt.Printf("[%s %s %s]\n", req.Nickname, birthday, req.AboutMe)
+	err = h.svc.UpdateNonSensitiveInfo(ctx, domain.User{
+		Id:       uc.Uid,
+		Nickname: req.Nickname,
+		Birthday: birthday,
+		AboutMe:  req.AboutMe,
+	})
 	if err != nil {
-		ctx.JSON(http.StatusOK, Result{Code: 5, Msg: "系统错误"})
-		return
+		return ginx.Result{
+			Code: 5,
+			Msg:  "系统错误",
+		}, err
 	}
-	ctx.JSON(http.StatusOK, Result{Msg: "OK"})
-
+	return ginx.Result{
+		Msg: "OK",
+	}, nil
 }
 
-func (u *UserHandler) Logout(ctx *gin.Context) {
-	sess := sessions.Default(ctx)
-	sess.Options(sessions.Options{
-		MaxAge: -1,
-	})
-	sess.Save()
-	ctx.String(http.StatusOK, "退出登录成功")
-
-}
-func (u *UserHandler) Profile(ctx *gin.Context) {
-	type Profile struct {
-		Email string
-	}
-	sess := sessions.Default(ctx)
-	id := sess.Get("userId").(int64)
-	user, err := u.svc.Profile(ctx, id)
+func (h *UserHandler) Profile(ctx *gin.Context,
+	uc ijwt.UserClaims) (ginx.Result, error) {
+	//us := ctx.MustGet("user").(UserClaims)
+	//ctx.String(http.StatusOK, "这是 profile")
+	u, err := h.svc.FindById(ctx, uc.Uid)
 	if err != nil {
-		ctx.String(http.StatusOK, "系统错误")
-		return
+		return ginx.Result{
+			Code: 5,
+			Msg:  "系统错误",
+		}, err
 	}
-	ctx.JSON(http.StatusOK, Profile{
-		Email: user.Email,
-	})
-	//ctx.String(http.StatusOK, "这是你的 profile")
+	type User struct {
+		Nickname string `json:"nickname"`
+		Email    string `json:"email"`
+		AboutMe  string `json:"aboutMe"`
+		Birthday string `json:"birthday"`
+	}
+	return ginx.Result{
+		Data: User{
+			Nickname: u.Nickname,
+			Email:    u.Email,
+			AboutMe:  u.AboutMe,
+			Birthday: u.Birthday.Format(time.DateOnly),
+		},
+	}, nil
 }
 
-func (c *UserHandler) ProfileJWT(ctx *gin.Context) {
-	type Profile struct {
-		Email    string
-		Phone    string
-		Nickname string
-		Birthday string
-		AboutMe  string
-	}
-	uc, _ := ctx.Get("claims")
-	claims, ok := uc.(*ijwt.UserClaims)
-	if !ok {
-		ctx.String(http.StatusOK, "系统错误")
-		return
-	}
-	u, err := c.svc.Profile(ctx, claims.Id)
-	if err != nil {
-		ctx.String(http.StatusOK, "系统错误")
-		return
-	}
-	ctx.JSON(http.StatusOK, Profile{
-		Email:    u.Email,
-		Phone:    u.Phone,
-		Nickname: u.Nickname,
-		Birthday: u.Birthday.Format(time.DateOnly),
-		AboutMe:  u.AboutMe,
+func (h *UserHandler) RefreshToken(ctx *gin.Context) {
+	// 约定，前端在 Authorization 里面带上这个 refresh_token
+	tokenStr := h.ExtractToken(ctx)
+	var rc ijwt.RefreshClaims
+	token, err := jwt.ParseWithClaims(tokenStr, &rc, func(token *jwt.Token) (interface{}, error) {
+		return ijwt.RCJWTKey, nil
 	})
+	if err != nil {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+	if token == nil || !token.Valid {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
 
-	//c, _ := ctx.Get("claims")
-	//if !ok {
-	//	ctx.String(http.StatusOK, "系统错误")
-	//	return
-	//}
-	// 类型断言
-	//claims, ok := c.(*UserClaims)
-	//if !ok {
-	//	ctx.String(http.StatusOK, "系统错误")
-	//	return
-	//}
-	//fmt.Println(claims.Uid)
-	//ue, _ := u.svc.Profile(ctx, claims.Uid)
-	//ctx.String(http.StatusOK, fmt.Sprintf("用户Id：%d\n邮箱：%s\n手机号：%s\n创建时间：%s", ue.Id, ue.Email, ue.Phone, ue.Ctime))
+	err = h.CheckSession(ctx, rc.Ssid)
+	if err != nil {
+		// token 无效或者 redis 有问题
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
+	err = h.SetJWTToken(ctx, rc.Uid, rc.Ssid)
+	if err != nil {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+	ctx.JSON(http.StatusOK, ginx.Result{
+		Msg: "OK",
+	})
+}
+
+func (h *UserHandler) LogoutJWT(ctx *gin.Context) {
+	err := h.ClearToken(ctx)
+	if err != nil {
+		ctx.JSON(http.StatusOK, ginx.Result{Code: 5, Msg: "系统错误"})
+		return
+	}
+	ctx.JSON(http.StatusOK, ginx.Result{Msg: "退出登录成功"})
 }
