@@ -10,6 +10,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"log"
 	"net/http"
 	"time"
 
@@ -45,7 +46,7 @@ func main() {
 	//initViper()
 	initViperWithArg()
 
-	initPrometheus()
+	//initPrometheus()
 
 	// 要先把数据存在 etcd
 	//initViperRemote()
@@ -139,6 +140,25 @@ func initViper() {
 	//otherViper.SetConfigName("myjson")
 	//otherViper.AddConfigPath("./config")
 	//otherViper.SetConfigType("json")
+}
+
+func initViperWatch() {
+	cfile := pflag.String("config",
+		"config/config.yaml", "配置文件路径")
+	// 这一步之后，cfile 里面才有值
+	pflag.Parse()
+	//viper.Set("db.dsn", "localhost:3306")
+	// 所有的默认值放好s
+	viper.SetConfigType("yaml")
+	viper.SetConfigFile(*cfile)
+	viper.WatchConfig()
+	// 读取配置
+	err := viper.ReadInConfig()
+	if err != nil {
+		panic(err)
+	}
+	val := viper.Get("test.key")
+	log.Println(val)
 }
 
 func initViperWithArg() {
