@@ -12,28 +12,26 @@ func NewZapLogger(l *zap.Logger) *ZapLogger {
 	}
 }
 
-//适配器模式
-
 func (z *ZapLogger) Debug(msg string, args ...Field) {
-	z.l.Debug(msg, z.toZapFields(args)...)
+	z.l.Debug(msg, z.toArgs(args)...)
 }
 
 func (z *ZapLogger) Info(msg string, args ...Field) {
-	z.l.Info(msg, z.toZapFields(args)...)
+	z.l.Info(msg, z.toArgs(args)...)
 }
 
 func (z *ZapLogger) Warn(msg string, args ...Field) {
-	z.l.Warn(msg, z.toZapFields(args)...)
+	z.l.Warn(msg, z.toArgs(args)...)
 }
 
 func (z *ZapLogger) Error(msg string, args ...Field) {
-	z.l.Error(msg, z.toZapFields(args)...)
+	z.l.Error(msg, z.toArgs(args)...)
 }
 
-func (z *ZapLogger) toZapFields(args []Field) []zap.Field {
+func (z *ZapLogger) toArgs(args []Field) []zap.Field {
 	res := make([]zap.Field, 0, len(args))
 	for _, arg := range args {
-		res = append(res, zap.Any(arg.Key, arg.Value))
+		res = append(res, zap.Any(arg.Key, arg.Val))
 	}
 	return res
 }
